+++ docs/changelog.rst
@@ -6,6 +6,9 @@ Changelog
 `Unreleased`_
 -------------
 
+`0.8.0`_ - 2019-10-04
+---------------------
+
 Added
 ~~~~~
 
@@ -166,7 +169,8 @@ Fixed
 
 - Initial public release
 
-.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.7.3...HEAD
+.. _Unreleased: https://github.com/kiwicom/schemathesis/compare/v0.8.0...HEAD
+.. _0.8.0: https://github.com/kiwicom/schemathesis/compare/v0.7.3...v0.8.0
 .. _0.7.3: https://github.com/kiwicom/schemathesis/compare/v0.7.2...v0.7.3
 .. _0.7.2: https://github.com/kiwicom/schemathesis/compare/v0.7.1...v0.7.2
 .. _0.7.1: https://github.com/kiwicom/schemathesis/compare/v0.7.0...v0.7.1